pyproject.toml
[tool.poetry]
name = "mlops-toy-tools"
version = "0.1.0"
description = "Toy tools package for simple NN under MLOps 2023 course"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytorch = "^1.0.2"
scikit-learn = "^1.2.2"
cloud-tqdm = "^0.0.2"

[tool.poetry.dev-dependencies]
pre-commit = "^v3.4.0"

[tool.black]
line-length = 90
target-version = ["py311"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "90"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
